+++ lib/ui/screens/security/security_login_biometrics_screen.dart
@@ -10,6 +10,11 @@ import 'package:mooncake/ui/ui.dart';
 /// setting up the account.
 class LoginWithBiometricsScreen extends StatelessWidget {
   final LocalAuthentication localAuth = Injector.get();
+  final bool backupPhrase;
+
+  LoginWithBiometricsScreen({
+    this.backupPhrase = false,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -24,42 +29,46 @@ class LoginWithBiometricsScreen extends StatelessWidget {
           builder: (context, state) {
             return Stack(
               children: [
-                Container(
-                  padding: EdgeInsets.all(16),
-                  child: ListView(
-                    children: [
-                      Text(
-                        PostsLocalizations.of(context)
-                            .securityLoginText
-                            .replaceAll("\n", ""),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        PostsLocalizations.of(context)
-                            .securityLoginWarning
-                            .replaceAll("\n", ""),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        PostsLocalizations.of(context)
-                            .securityLoginBiometrics
-                            .replaceAll("\n", ""),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(height: 16),
-                      if (!state.showMnemonic)
-                        PrimaryButton(
-                          onPressed: () => _enableButtonClicked(context),
-                          child:
-                              Text(PostsLocalizations.of(context).viewMnemonic),
+                state.showMnemonic
+                    ? MnemonicVisualizer(
+                        mnemonic: state.mnemonic,
+                        allowExport: backupPhrase ? false : true,
+                        backupPhrase: backupPhrase,
+                      )
+                    : Container(
+                        padding: EdgeInsets.all(16),
+                        child: ListView(
+                          children: [
+                            Text(
+                              PostsLocalizations.of(context)
+                                  .securityLoginText
+                                  .replaceAll("\n", ""),
+                              textAlign: TextAlign.center,
+                            ),
+                            const SizedBox(height: 16),
+                            Text(
+                              PostsLocalizations.of(context)
+                                  .securityLoginWarning
+                                  .replaceAll("\n", ""),
+                              textAlign: TextAlign.center,
+                            ),
+                            const SizedBox(height: 16),
+                            Text(
+                              PostsLocalizations.of(context)
+                                  .securityLoginBiometrics
+                                  .replaceAll("\n", ""),
+                              textAlign: TextAlign.center,
+                            ),
+                            const SizedBox(height: 16),
+                            if (!state.showMnemonic)
+                              PrimaryButton(
+                                onPressed: () => _enableButtonClicked(context),
+                                child: Text(PostsLocalizations.of(context)
+                                    .viewMnemonic),
+                              ),
+                          ],
                         ),
-                      if (state.showMnemonic)
-                        MnemonicVisualizer(mnemonic: state.mnemonic),
-                    ],
-                  ),
-                ),
+                      ),
 
                 // Exporting popup
                 if (state is ExportingMnemonic) ExportMnemonicPopup(),