+++ lib/ui/widgets/buttons/export.dart
@@ -2,3 +2,4 @@ export 'check_box_button.dart';
 export 'primary_light_button.dart';
 export 'primary_button.dart';
 export 'secondary_button.dart';
+export 'secondary_invert_button.dart';