+++ lib/ui/widgets/mnemonic/mnemonic_visualizer.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:mooncake/ui/ui.dart';
 
@@ -9,41 +8,35 @@ import 'mnemonic_item.dart';
 /// inserted himself or created from scratch.
 class MnemonicVisualizer extends StatelessWidget {
   final List<String> mnemonic;
-
+  final bool allowExport;
+  final bool backupPhrase;
   const MnemonicVisualizer({
     Key key,
     @required this.mnemonic,
+    this.allowExport = true,
+    this.backupPhrase = false,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<MnemonicBloc, MnemonicState>(
-      builder: (context, state) {
-        return Column(
-          children: [
-            StaggeredGridView.countBuilder(
-              crossAxisSpacing: 10,
-              mainAxisSpacing: 10,
-              crossAxisCount: 3,
-              itemCount: 24,
-              shrinkWrap: true,
-              physics: NeverScrollableScrollPhysics(),
-              staggeredTileBuilder: (index) => StaggeredTile.fit(1),
-              itemBuilder: (BuildContext context, int index) {
-                return MnemonicItem(index: index + 1, word: mnemonic[index]);
-              },
-            ),
-            PrimaryButton(
-              child: Text(PostsLocalizations.of(context).exportMnemonic),
-              onPressed: () => _openExportPopup(context),
-            )
-          ],
-        );
-      },
+    return Column(
+      children: [
+        if (backupPhrase) BackupMnemonicDetails(),
+        StaggeredGridView.countBuilder(
+          crossAxisSpacing: 10,
+          mainAxisSpacing: 10,
+          crossAxisCount: 3,
+          itemCount: 24,
+          shrinkWrap: true,
+          physics: NeverScrollableScrollPhysics(),
+          staggeredTileBuilder: (index) => StaggeredTile.fit(1),
+          itemBuilder: (BuildContext context, int index) {
+            return MnemonicItem(index: index + 1, word: mnemonic[index]);
+          },
+        ),
+        if (allowExport) ExportMnemonicButton(),
+        if (backupPhrase) BackupMnemonicButton()
+      ],
     );
-  }
-
-  void _openExportPopup(BuildContext context) {
-    BlocProvider.of<MnemonicBloc>(context).add(ShowExportPopup());
   }
 }