+++ lib/usecases/dependency_injection.dart
@@ -101,7 +101,9 @@ class UseCaseModule implements Module {
       ..bindFactory((injector, params) => GetSettingUseCase(
             settingsRepository: injector.get(),
           ))
-
+      ..bindFactory((injector, params) => WatchSettingUseCase(
+            settingsRepository: injector.get(),
+          ))
       // Users use cases
       ..bindFactory((injector, params) => BlockUserUseCase(
             usersRepository: injector.get(),